requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv
motor==3.4.0
pymongo==4.8.0
openai==1.61.1